+++ sdk/resourcemanager/mediaservices/armmediaservices/operationstatuses_client.go
@@ -23,7 +23,7 @@ import (
 	"strings"
 )
 
-// OperationStatusesClient contains the methods for the OperationStatuses group.
+// OperationStatusesClient contains the methods for the MediaServicesOperationStatuses group.
 // Don't use this type directly, use NewOperationStatusesClient() instead.
 type OperationStatusesClient struct {
 	host           string
@@ -55,17 +55,14 @@ func NewOperationStatusesClient(subscriptionID string, credential azcore.TokenCr
 	return client, nil
 }
 
-// Get - Get asset track operation status.
+// Get - Get media service operation status.
 // If the operation fails it returns an *azcore.ResponseError type.
 // Generated from API version 2021-11-01
-// resourceGroupName - The name of the resource group within the Azure subscription.
-// accountName - The Media Services account name.
-// assetName - The Asset name.
-// trackName - The Asset Track name.
-// operationID - Operation Id.
+// locationName - Location name.
+// operationID - Operation ID.
 // options - OperationStatusesClientGetOptions contains the optional parameters for the OperationStatusesClient.Get method.
-func (client *OperationStatusesClient) Get(ctx context.Context, resourceGroupName string, accountName string, assetName string, trackName string, operationID string, options *OperationStatusesClientGetOptions) (OperationStatusesClientGetResponse, error) {
-	req, err := client.getCreateRequest(ctx, resourceGroupName, accountName, assetName, trackName, operationID, options)
+func (client *OperationStatusesClient) Get(ctx context.Context, locationName string, operationID string, options *OperationStatusesClientGetOptions) (OperationStatusesClientGetResponse, error) {
+	req, err := client.getCreateRequest(ctx, locationName, operationID, options)
 	if err != nil {
 		return OperationStatusesClientGetResponse{}, err
 	}
@@ -80,28 +77,16 @@ func (client *OperationStatusesClient) Get(ctx context.Context, resourceGroupNam
 }
 
 // getCreateRequest creates the Get request.
-func (client *OperationStatusesClient) getCreateRequest(ctx context.Context, resourceGroupName string, accountName string, assetName string, trackName string, operationID string, options *OperationStatusesClientGetOptions) (*policy.Request, error) {
-	urlPath := "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Media/mediaServices/{accountName}/assets/{assetName}/tracks/{trackName}/operationStatuses/{operationId}"
+func (client *OperationStatusesClient) getCreateRequest(ctx context.Context, locationName string, operationID string, options *OperationStatusesClientGetOptions) (*policy.Request, error) {
+	urlPath := "/subscriptions/{subscriptionId}/providers/Microsoft.Media/locations/{locationName}/mediaServicesOperationStatuses/{operationId}"
 	if client.subscriptionID == "" {
 		return nil, errors.New("parameter client.subscriptionID cannot be empty")
 	}
 	urlPath = strings.ReplaceAll(urlPath, "{subscriptionId}", url.PathEscape(client.subscriptionID))
-	if resourceGroupName == "" {
-		return nil, errors.New("parameter resourceGroupName cannot be empty")
+	if locationName == "" {
+		return nil, errors.New("parameter locationName cannot be empty")
 	}
-	urlPath = strings.ReplaceAll(urlPath, "{resourceGroupName}", url.PathEscape(resourceGroupName))
-	if accountName == "" {
-		return nil, errors.New("parameter accountName cannot be empty")
-	}
-	urlPath = strings.ReplaceAll(urlPath, "{accountName}", url.PathEscape(accountName))
-	if assetName == "" {
-		return nil, errors.New("parameter assetName cannot be empty")
-	}
-	urlPath = strings.ReplaceAll(urlPath, "{assetName}", url.PathEscape(assetName))
-	if trackName == "" {
-		return nil, errors.New("parameter trackName cannot be empty")
-	}
-	urlPath = strings.ReplaceAll(urlPath, "{trackName}", url.PathEscape(trackName))
+	urlPath = strings.ReplaceAll(urlPath, "{locationName}", url.PathEscape(locationName))
 	if operationID == "" {
 		return nil, errors.New("parameter operationID cannot be empty")
 	}
@@ -120,7 +105,7 @@ func (client *OperationStatusesClient) getCreateRequest(ctx context.Context, res
 // getHandleResponse handles the Get response.
 func (client *OperationStatusesClient) getHandleResponse(resp *http.Response) (OperationStatusesClientGetResponse, error) {
 	result := OperationStatusesClientGetResponse{}
-	if err := runtime.UnmarshalAsJSON(resp, &result.AssetTrackOperationStatus); err != nil {
+	if err := runtime.UnmarshalAsJSON(resp, &result.MediaServiceOperationStatus); err != nil {
 		return OperationStatusesClientGetResponse{}, err
 	}
 	return result, nil